refactor: Add redirect for publishing action
The publishing action assumes that main in package.json points to the
entrypoint of the action. Since v1.2.0, however, that is no longer the
case. Rather than try a new system (or dive deep into the internals of
npm and what every directive means), we'll trick the action into uploading
the correct script by modifying the package.json file right before upload.
Add that hack.